+++ lib/features/cart/domain/repository/stripe_repository.dart
@@ -3,6 +3,6 @@ import 'package:food_hub/features/cart/data/model/payment_model.dart';
 
 abstract class StripeRepository {
   Future<Either<String, dynamic>> createPaymentIntent(
-      int amount, String currency);
+      double amount, String currency);
   Future<Either<String, dynamic>> makePayment(PaymentModel paymentModel);
 }
